Refactor query printing
