+++ src/selectionSetToQuery.js
@@ -2,48 +2,87 @@ import { print } from 'graphql/language';
 
 export function selectionSetToNodeQueryDefinition({
   id,
+  typeName,
   selectionSet,
 }) {
   const queryDocumentAst = {
     kind: 'Document',
     definitions: [
-      {
-        kind: 'OperationDefinition',
-        operation: 'query',
-        name: null,
-        variableDefinitions: null,
-        directives: [],
-        selectionSet: {
-          kind: 'SelectionSet',
-          selections: [
-            {
-              kind: 'Field',
-              alias: null,
-              name: {
-                kind: 'Name',
-                value: 'node',
-              },
-              arguments: [
-                {
-                  kind: 'Argument',
-                  name: {
-                    kind: 'Name',
-                    value: 'id',
-                  },
-                  value: {
-                    kind: 'StringValue',
-                    value: id,
-                  },
-                },
-              ],
-              directives: [],
-              selectionSet,
-            },
-          ],
-        },
-      },
+      nodeQueryDefinition({
+        id,
+        typeName,
+        selectionSet,
+      }),
     ],
   };
 
   return print(queryDocumentAst);
+}
+
+function nodeQueryDefinition({
+  id,
+  typeName,
+  selectionSet,
+}) {
+  return {
+    kind: 'OperationDefinition',
+    operation: 'query',
+    name: null,
+    variableDefinitions: null,
+    directives: [],
+    selectionSet: {
+      kind: 'SelectionSet',
+      selections: [
+        {
+          kind: 'Field',
+          alias: null,
+          name: {
+            kind: 'Name',
+            value: 'node',
+          },
+          arguments: [
+            {
+              kind: 'Argument',
+              name: {
+                kind: 'Name',
+                value: 'id',
+              },
+              value: {
+                kind: 'StringValue',
+                value: id,
+              },
+            },
+          ],
+          directives: [],
+          selectionSet: {
+            kind: 'SelectionSet',
+            selections: [
+              inlineFragmentSelection({
+                typeName,
+                selectionSet,
+              }),
+            ],
+          },
+        },
+      ],
+    },
+  };
+}
+
+function inlineFragmentSelection({
+  typeName,
+  selectionSet,
+}) {
+  return {
+    kind: 'InlineFragment',
+    typeCondition: {
+      kind: 'NamedType',
+      name: {
+        kind: 'Name',
+        value: typeName,
+      },
+    },
+    directives: [],
+    selectionSet,
+  };
 }
