+++ test/tests.js
@@ -2,3 +2,4 @@ import './writeToStore';
 import './readFromStore';
 import './roundtrip';
 import './diffAgainstStore';
+import './queryPrinting';
